latitude and longitude display in ST9 and ST12 of CLAS. fixed CLAS decoder bug in concatinating data part.
